refactored previous code based on the new style guide
