+++ src/components/Donations/donation-amount.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import Image from "next/image";
+
 import BequestIcon7 from "~/images/donations/bequests/BequestIcon7.png";
 import BequestIcon8 from "~/images/donations/bequests/BequestIcon8.png";
 
@@ -9,7 +10,11 @@ interface ItemProps {
   description: string;
 }
 
-const DonationAmount: React.FC<ItemProps> = ({ pos, amount, description }) => {
+export default function DonationAmount({
+  pos,
+  amount,
+  description
+}: ItemProps) {
   if (pos) {
     return (
       <div className="grid grid-cols-2 place-items-center">
@@ -49,5 +54,4 @@ const DonationAmount: React.FC<ItemProps> = ({ pos, amount, description }) => {
       </div>
     );
   }
-};
-export default DonationAmount;
+}
