+++ src/components/Donations/donation-approach.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 
-const DonationApproach: React.FC<{ order: number; text: string }> = ({
-  order,
-  text
-}) => {
+interface ItemProps {
+  order: number;
+  text: string;
+}
+
+export default function DonationApproach({ order, text }: ItemProps) {
   return (
     <div className="flex items-center justify-items-start">
       <div className="mx-5 my-1 text-4xl font-bold text-[--starick-orange]">
@@ -15,5 +17,4 @@ const DonationApproach: React.FC<{ order: number; text: string }> = ({
       />
     </div>
   );
-};
-export default DonationApproach;
+}
