+++ src/components/Donations/donation-type.tsx
@@ -1,15 +1,12 @@
 import React from "react";
 import Image, { StaticImageData } from "next/image";
-import BequestIcon7 from "~/images/donations/bequests/BequestIcon7.png";
-import BequestIcon8 from "~/images/donations/bequests/BequestIcon8.png";
-import BequestIcon2 from "~/images/donations/bequests/BequestIcon2.png";
 
 interface ItemProps {
   image: StaticImageData;
   text: string;
 }
 
-const DonationTitle: React.FC<ItemProps> = ({ image, text }) => {
+export default function DonationTitle({ image, text }: ItemProps) {
   return (
     <div className="flex flex-col">
       <Image
@@ -23,5 +20,4 @@ const DonationTitle: React.FC<ItemProps> = ({ image, text }) => {
       <p className="font-bold italic text-[--starick-brown]">{text}</p>
     </div>
   );
-};
-export default DonationTitle;
+}
