+++ src/components/Home/core-service-item.tsx
@@ -6,12 +6,11 @@ interface ItemProps {
   text: string;
 }
 
-const CoreServiceItem: React.FC<ItemProps> = ({ img, text }) => {
+export default function CoreServiceItem({ img, text }: ItemProps) {
   return (
     <div className="flex items-center justify-items-start">
       <Image src={img} alt="description" width={150} height={150} />
       <p className="ml-20 font-bold text-[--starick-white]">{text}</p>
     </div>
   );
-};
-export default CoreServiceItem;
+}
